requirements.txt
flask>=3.0.0
docling>=1.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
flask-cors>=4.0.0

